Revert "Split fuelHandlers.py into multiple scripts"
This reverts commit 5aa37bb0d4ec3054705fefb6e099490505b5577d.
